Initial commit.
This is the first commit in a project meant to provide simple resouce locks for
an erlang cluster. These are volatile locks and not intended to be distributed
over more than a reliable-ish network cluster (unless the user makes contingency
plans otherwise).

All of the above is aspirational. This commit is bits and pieces.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>